refactor: move the search-region grid math into acq4/experiment
The serpentine tile grid is engine logic, needed by the cell producer, and
its tests only ran where acq4_automation is installed because they lived
under AutomationDebug/tests. SurveyRegion re-exports from the new home.
